Fixed failing test on wahba

+++ pyquat/wahba/qmethod.py
@@ -22,12 +22,12 @@ import scipy.linalg as spl
 
 if TYPE_CHECKING:
     import pyquat._pyquat as _pq
+    from wahba import attitude_profile_matrix
+    from wahba import davenport_matrix
 else:
     import _pyquat as _pq
-from wahba import attitude_profile_matrix
-from wahba import davenport_matrix
-
-
+    from pyquat.wahba import attitude_profile_matrix
+    from pyquat.wahba import davenport_matrix
 
 def qekf_measurement_model(
         T: np.ndarray[Any, np.dtype[np.float64]],
@@ -78,7 +78,7 @@ def quest_measurement_covariance(
         A 3x3 covariance matrix.
     """
     
-    return ((np.identity(3) - vector[0:3].reshape((3, 1)).dot(vector[0:3].reshape((1,3)))) * sigma).astype(np.float64)
+    return (np.identity(3) - vector[0:3].reshape((3, 1)).dot(vector[0:3].reshape((1,3)))) * sigma
 
 
 def qekf_measurement_covariance(
